Correções no ranking e finalização do quiz
