+++ components/quiz/IniciarQuiz.tsx
@@ -44,7 +44,7 @@ const DadosFunci = ({ matricula, setMatricula }: any) => {
         Digite a sua matrícula
       </div>
       <input
-        className="bg-[#292929] w-/12 shadow appearance-none border border-ddBlack rounded text-white p-4 text-lg leading-tight focus:outline-none focus:shadow-outline"
+        className="bg-[#292929] w-/12 shadow appearance-none border border-ddBlack rounded text-white p-4 text-lg leading-tight focus:outline-none focus:shadow-outline placeholder-ddLightGrey"
         onChange={(e) => setMatricula(e.target.value)}
         value={matricula}
       />
@@ -102,7 +102,7 @@ export default function IniciarQuiz({
       },
     };
     try {
-      const response = await axios.post("http://localhost:3000/quiz/1", {
+      const response = await axios.post("http://localhost:8080/participacao/quiz/1", {
         email,
         nome,
         matricula,
@@ -133,6 +133,17 @@ export default function IniciarQuiz({
     }
 
     if (tipoRespondente === TIPO_RESPONDENTE_EXTERNO && (!nome || !email)) {
+      return false;
+    }
+
+
+    const regexMatricula = /^(F|C)\d{7}$/;
+    if(matricula && !regexMatricula.test(matricula)){
+      return false
+    }
+
+    const regexEmail = /[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/;
+    if(email && !regexEmail.test(email)){
       return false;
     }
