+++ pages/ranking.module.scss
